community/dai-complex-workflow/scripts/utils/requirements.txt
google-cloud-service-usage==1.2.1
google-cloud-documentai==1.0.0